Address review cleanup: drop dead mapToRange overload, fix Map fallback normalization, align Variant NULL hash
Cleanup pass on the ColumnsScatter/computeHashInto branch:

- Remove the dead `UInt64` `mapToRange` overload and `mapToRangeImplU64`
  (no production caller; both transforms use the `UInt32` pids path) and
  switch `benchmark_compute_hash_into` to `UInt32` pids.
- `deepNormalizeForFallback`: add a `Map` branch that deep-normalizes the
  nested `Array(Tuple(K, V))` and recreates `ColumnMap`, so a batched flush
  mixing full-nested and `ColumnSparse`/`ColumnReplicated`-nested `Map`
  sources no longer hits the invalid `insertRangeFrom`/`assert_cast` path.
  Add `FallbackMapMixedNestedRepresentation` regression.
- `ColumnVariant::computeHashInto`: use `WEAK_HASH32_INITIAL_VALUE` for NULL
  rows to match `ColumnNullable`/`IColumnDummy`, and fix the comment.
- `shard_by_hash_input_batch_bytes` doc: note the internal chunk-count safety
  cap (a flush can happen before the byte budget for const-heavy chunks).
- Revert unrelated churn in `benchmark_column_insert_many_from`; drop the
  unused `<algorithm>` include in `JoinUtils`; trim stale/dev-history comments
  in `ColumnsScatter`, `SettingsChangesHistory`, and the SQL tests.
- Rename `04278_04101_sharded_aggregation_batched_input` to a single
  sequential prefix `04317_`.

+++ src/Columns/benchmarks/benchmark_compute_hash_into.cpp
@@ -122,12 +122,12 @@ void BM_ComputeHashInto(benchmark::State & state, const std::vector<ColumnPtr> &
     state.counters["K"] = static_cast<double>(ncols);
 }
 
-/// ─── End-to-end hash → selector benchmark ────────────────────────────────
+/// ─── End-to-end hash → partition-id benchmark ────────────────────────────────
 ///
-/// Captures the full pipeline exactly as executed by
-/// BufferedShardByHashTransform::generateOutputChunks:
-/// seed hash_buffer with WEAK_HASH32_INITIAL_VALUE, chain computeHashInto with
-/// initial=false over every key column (0 allocations), then mapToRange SIMD.
+/// Mirrors the hash-production step of the hash-partitioning consumers
+/// (BufferedShardByHashTransform, ScatterByPartitionTransform): seed hash_buffer
+/// with WEAK_HASH32_INITIAL_VALUE, chain computeHashInto with initial=false over
+/// every key column (0 allocations), then mapToRange SIMD to UInt32 pids.
 ///
 /// The per-iteration std::fill is part of the consumer hot path (the buffer is
 /// re-seeded for every block), so it is measured here rather than hoisted out.
@@ -139,7 +139,7 @@ void BM_HashAndFastrange_New(benchmark::State & state, const std::vector<ColumnP
     const size_t n = cols[0]->size();
     const size_t ncols = cols.size();
     PaddedPODArray<UInt32> hash_buf(n);
-    PaddedPODArray<UInt64> pids(n); // mapToRange writes a UInt64 selector
+    PaddedPODArray<UInt32> pids(n); // mapToRange writes UInt32 partition ids
     const UInt32 p32 = static_cast<UInt32>(num_shards);
     for (auto _ : state)
     {
